+++ BE/src/main/java/com/g3/feedbackApp/Controllers/UserController.java
@@ -1,10 +1,7 @@
 package com.g3.feedbackApp.Controllers;
 
-import com.g3.feedbackApp.DataSources.FakeDataSourceUser;
 import com.g3.feedbackApp.Models.UserModel;
 import com.g3.feedbackApp.Services.Interfaces.IUserService;
-import com.g3.feedbackApp.Services.ReviewerService;
-import com.g3.feedbackApp.Services.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -26,7 +23,7 @@ public class UserController {
 
     @GetMapping("{studentNr}")
     public ResponseEntity<UserModel> getUserPath(@PathVariable(value = "studentNr") int studentNr) {
-        UserModel userModel = userService.getUserByStudentNr(studentNr);
+        UserModel userModel = userService.getUserByStudentNr((long)studentNr);
 
         if(userModel != null) {
             return ResponseEntity.ok().body(userModel);
@@ -60,7 +57,7 @@ public class UserController {
 
     @DeleteMapping("{studentNr}")
     public ResponseEntity deletePost(@PathVariable int studentNr) {
-        userService.deleteUserModel(studentNr);
+        userService.deleteUserModel((long)studentNr);
         return ResponseEntity.ok().build();
 
     }
@@ -99,7 +96,7 @@ public class UserController {
                                                    @RequestParam("email") String email
                                                    ) {
 
-        UserModel userModel = userService.getUserByStudentNr(studentNr);
+        UserModel userModel = userService.getUserByStudentNr((long)studentNr);
 
         if (userModel == null){
             return new ResponseEntity("Please provide a valid student number.",HttpStatus.NOT_FOUND);
