+++ BE/src/main/java/com/g3/feedbackApp/Repository/ICommentRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ICommentRepository extends JpaRepository<CommentModel, Integer> {
+public interface ICommentRepository extends JpaRepository<CommentModel, Long> {
 
-    CommentModel getFirstByCommentId(int id);
-    List<CommentModel> getCommentModelsByVersionId(int id);
+    CommentModel getFirstByCommentId(Long id);
+    List<CommentModel> getCommentModelsByVersionId(Long id);
 }
