+++ BE/src/main/java/com/g3/feedbackApp/Services/Interfaces/IUserService.java
@@ -5,10 +5,10 @@ import com.g3.feedbackApp.Models.UserModel;
 import java.util.List;
 
 public interface IUserService {
-    public UserModel getUserByStudentNr(int studentNr);
+    public UserModel getUserByStudentNr(Long studentNr);
     public List<UserModel> getUserModels();
     public UserModel getUserByEmail(String email);
-    public boolean deleteUserModel(int studentNr);
+    public boolean deleteUserModel(Long studentNr);
     public boolean addUserModel(UserModel userModel);
     public boolean updateUserModel(UserModel userModel);
 }
